+++ plugins/forwarder/grpc/envoymetricsv3/forwarder.go
@@ -33,7 +33,10 @@ import (
 	"github.com/apache/skywalking-satellite/internal/satellite/event"
 )
 
-const Name = "envoy-metrics-v3-grpc-forwarder"
+const (
+	Name     = "envoy-metrics-v3-grpc-forwarder"
+	ShowName = "Envoy Metrics v3 GRPC Forwarder"
+)
 
 type Forwarder struct {
 	config.CommonFields
@@ -42,6 +45,10 @@ type Forwarder struct {
 
 func (f *Forwarder) Name() string {
 	return Name
+}
+
+func (f *Forwarder) ShowName() string {
+	return ShowName
 }
 
 func (f *Forwarder) Description() string {