+++ plugins/forwarder/grpc/nativejvm/sync_forwarder.go
@@ -31,7 +31,10 @@ import (
 	"github.com/apache/skywalking-satellite/internal/satellite/event"
 )
 
-const Name = "nativejvm-grpc-forwarder"
+const (
+	Name     = "nativejvm-grpc-forwarder"
+	ShowName = "Native JVM GRPC Forwarder"
+)
 
 type Forwarder struct {
 	config.CommonFields
@@ -41,6 +44,10 @@ type Forwarder struct {
 
 func (f *Forwarder) Name() string {
 	return Name
+}
+
+func (f *Forwarder) ShowName() string {
+	return ShowName
 }
 
 func (f *Forwarder) Description() string {