+++ plugins/queue/memory/queue.go
@@ -29,8 +29,13 @@ import (
 )
 
 const (
-	Name = "memory-queue"
+	Name     = "memory-queue"
+	ShowName = "Memory Queue"
 )
+
+var DefaultOffset = &event.Offset{
+	Position: "",
+}
 
 type Queue struct {
 	config.CommonFields
@@ -43,6 +48,10 @@ type Queue struct {
 
 func (q *Queue) Name() string {
 	return Name
+}
+
+func (q *Queue) ShowName() string {
+	return ShowName
 }
 
 func (q *Queue) Description() string {
@@ -77,7 +86,7 @@ func (q *Queue) Dequeue() (*api.SequenceEvent, error) {
 	}
 	return &api.SequenceEvent{
 		Event:  element.(*v1.SniffData),
-		Offset: "no_offset_in_memory_queue",
+		Offset: DefaultOffset,
 	}, nil
 }
 
@@ -85,5 +94,17 @@ func (q *Queue) Close() error {
 	return nil
 }
 
-func (q *Queue) Ack(_ event.Offset) {
+func (q *Queue) Ack(_ *event.Offset) {
+}
+
+func (q *Queue) TotalSize() int64 {
+	return int64(q.EventBufferSize)
+}
+
+func (q *Queue) UsedCount() int64 {
+	return int64(q.buffer.GetLen())
+}
+
+func (q *Queue) IsFull() bool {
+	return q.UsedCount() >= q.TotalSize()
 }