+++ plugins/server/prometheus/prometheus.go
@@ -28,7 +28,10 @@ import (
 	"github.com/apache/skywalking-satellite/internal/satellite/telemetry"
 )
 
-const Name = "prometheus-server"
+const (
+	Name     = "prometheus-server"
+	ShowName = "Prometheus Server"
+)
 
 type Server struct {
 	config.CommonFields
@@ -40,6 +43,10 @@ type Server struct {
 
 func (s *Server) Name() string {
 	return Name
+}
+
+func (s *Server) ShowName() string {
+	return ShowName
 }
 
 func (s *Server) Description() string {